backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "barber-appointment-website"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "werkzeug (>=3.1.4,<4.0.0)",
    "python-jose (>=3.3.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "taskipy (>=1.14.1,<2.0.0) ; python_version >= \"3.14\" and python_version < \"4.0\"",
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.14.8,<0.15.0)"
]

[tool.taskipy.tasks]
test = "pytest"
format = "ruff format ."
